homebrew: add ClassiCube Proof Of Concept
A really early stage PoC of ClassiCube. Currently builds as a bin
and won't work on our devices. Biggest problem will be the lack of
any OpenGL and GPU support. Otherwise it could be a nice demo.

Stashing the port here for the time being, someone more competent
could theoretically build a working thing out of this one day.